+++ dex/engine/src/main/java/org/dependencytrack/dex/engine/persistence/command/CreateWorkflowRunCommand.java
@@ -31,7 +31,7 @@ public record CreateWorkflowRunCommand(
         int workflowVersion,
         String queueName,
         @Nullable String concurrencyGroupId,
-        int priority,
+        short priority,
         @Nullable Map<String, String> labels,
         Instant createdAt) {
 }
